Hours Entered Report Month Summary Report

+++ TimeTracker/TimeTracker/Shared/Models/Reports/MonthSummaryByUserDto.cs
@@ -7,6 +7,6 @@
         public string ProjectTitle { get; set; }
         public int ReferenceProcesses { get; set; }
         public string UserName { get; set; }
-        public float Hours { get; set; }
+        public decimal Hours { get; set; }
     }
 }
